+++ src/Locator/Exception/ExceptionInterface.php
@@ -1,0 +1,10 @@
+<?php
+declare(strict_types=1);
+
+namespace Remorhaz\JSON\Pointer\Locator\Exception;
+
+use Remorhaz\JSON\Pointer\Exception\ExceptionInterface as PointerExceptionInterface;
+
+interface ExceptionInterface extends PointerExceptionInterface
+{
+}