+++ src/Locator/LocatorBuilderInterface.php
@@ -1,9 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Remorhaz\JSON\Pointer\Parser;
-
-use Remorhaz\JSON\Pointer\Locator\LocatorInterface;
+namespace Remorhaz\JSON\Pointer\Locator;
 
 interface LocatorBuilderInterface
 {